+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version UNRELEASED
+# Version 22.11.0
 
 * Added The Second Helvetic Confession
 * Added The Augsburg Confession